КD-18: CRUD для опросов

+++ backend/src/main/java/ru/hh/kakdela_v2/service/AnswerOptionService.java
@@ -1,10 +1,16 @@
 package ru.hh.kakdela_v2.service;
 
+import org.springframework.stereotype.Service;
+import lombok.RequiredArgsConstructor;
 import ru.hh.kakdela_v2.dao.AnswerOptionDao;
 import ru.hh.kakdela_v2.dao.QuestionDao;
 import ru.hh.kakdela_v2.dto.answer_option.AnswerOptionCreateDto;
 import ru.hh.kakdela_v2.dto.answer_option.AnswerOptionResponseDto;
+import ru.hh.kakdela_v2.dto.answer_option.AnswerOptionUpdateDto;
+import ru.hh.kakdela_v2.dto.question.QuestionResponseDto;
+import ru.hh.kakdela_v2.dto.question.QuestionUpdateDto;
 import ru.hh.kakdela_v2.model.AnswerOption;
+import ru.hh.kakdela_v2.model.Permission.SurveyRole;
 import ru.hh.kakdela_v2.model.Question;
 import ru.hh.kakdela_v2.util.TransactionHelper;
 
@@ -14,17 +20,14 @@ import org.springframework.web.server.ResponseStatusException;
 import java.util.List;
 import java.util.UUID;
 
+@Service
+@RequiredArgsConstructor
 public class AnswerOptionService {
 
   private final AnswerOptionDao answerOptionDao;
+  private final PermissionService permissionService;
   private final QuestionDao questionDao;
   private final TransactionHelper transactionHelper;
-
-  public AnswerOptionService(AnswerOptionDao answerOptionDao, QuestionDao questionDao, TransactionHelper transactionHelper) {
-    this.answerOptionDao = answerOptionDao;
-    this.questionDao = questionDao;
-    this.transactionHelper = transactionHelper;
-  }
 
   public List<AnswerOptionResponseDto> getAllByQuestionId(UUID questionId) {
     return transactionHelper.inTransaction(() ->
@@ -34,10 +37,12 @@ public class AnswerOptionService {
     );
   }
 
-  public AnswerOptionResponseDto create(UUID questionId, AnswerOptionCreateDto dto) {
+  public AnswerOptionResponseDto create(UUID questionId, AnswerOptionCreateDto dto, UUID accountId) {
     return transactionHelper.inTransaction(() -> {
       Question question = questionDao.findById(questionId)
               .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Вопрос не найден: " + questionId));
+
+      permissionService.checkAccess(question.getSurveyPage().getSurvey().getId(), accountId, SurveyRole.EDITOR);        
 
       AnswerOption option = AnswerOption.builder()
               .question(question)
@@ -50,8 +55,27 @@ public class AnswerOptionService {
     });
   }
 
-  public void delete(UUID id) {
+  public AnswerOptionResponseDto update(UUID id, AnswerOptionUpdateDto dto, UUID accountId) {
+    return transactionHelper.inTransaction(() -> {
+      AnswerOption option = answerOptionDao.findById(id)
+              .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Вопрос не найден: " + id));
+
+      permissionService.checkAccess(option.getQuestion().getSurveyPage().getSurvey().getId(), accountId, SurveyRole.EDITOR);
+
+      if (dto.getSerialNumber() != null) option.setSerialNumber(dto.getSerialNumber());
+      if (dto.getAnswerOptionText() != null) option.setAnswerOptionText(dto.getAnswerOptionText());
+      answerOptionDao.update(option);
+      return new AnswerOptionResponseDto(option);
+    });
+  }
+
+  public void delete(UUID id, UUID accountId) {
     transactionHelper.inTransaction(() -> {
+      AnswerOption option = answerOptionDao.findById(id)
+                    .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Вариант ответа не найден: " + id));
+            
+            
+      permissionService.checkAccess(option.getQuestion().getSurveyPage().getSurvey().getId(), accountId, SurveyRole.EDITOR);
       answerOptionDao.delete(id);
     });
   }
